feat/refactor([back]blog-service): Post CRUD features
1. Code issued when posting
2. Create an upload folder with the code issued when registering an image
3. Apply Query DSL
4. Service logic supplementation

+++ backend/src/main/java/com/blog/notice/domain/Post.java
@@ -18,7 +18,7 @@ import lombok.Setter;
 public class Post extends BaseEntity {
 
     @Setter
-    private long boardId;
+    private int boardId;
     @Column(unique = true, nullable = false)
     private String code;
     private String title;
@@ -29,7 +29,7 @@ public class Post extends BaseEntity {
     private int views;
     private String tag;
 
-    public Post(long boardId, String code, String title, String subTitle, String author, String modifier
+    public Post(int boardId, String code, String title, String subTitle, String author, String modifier
             , int status, String tag) {
         this.boardId = boardId;
         this.code = code;
@@ -41,7 +41,7 @@ public class Post extends BaseEntity {
         this.tag = tag;
     }
 
-    public Post(long boardId, String code, String title, String author, int status) {
+    public Post(int boardId, String code, String title, String author, int status) {
         this.boardId = boardId;
         this.code = code;
         this.title = title;
@@ -54,10 +54,9 @@ public class Post extends BaseEntity {
                 request.getAuthor(), request.getModifier(), request.getStatus(), request.getTag());
     }
 
-    public static Post of(long boardId, String author) {
+    public static Post of(int boardId, String author) {
         return new Post(boardId, CodeGenerator.generateWithPrefix("POST"), "Temp Title", author, 1);
     }
-
 
     @Getter
     public enum PostStatus {
